feat(packages/rspack): `resolvedConfig`
