+++ .gitignore
@@ -108,6 +108,7 @@ out
 # Nuxt.js build / generate output
 .nuxt
 dist
+lib/
 
 # Gatsby files
 .cache/
