+++ packages/rspack/src/config/context.ts
@@ -1,0 +1,3 @@
+export type Context = string;
+
+export type ResolvedContext = string;
