+++ packages/rspack/src/config/define.ts
@@ -1,0 +1,3 @@
+export type Define = Record<string, string>;
+
+export type ResolvedDefine = Record<string, string>;
