+++ packages/rspack/src/config/entry.ts
@@ -1,0 +1,3 @@
+export type Entry = Record<string, string>;
+
+export type ResolvedEntry = Record<string, string>;
