+++ packages/rspack/src/config/external.ts
@@ -1,0 +1,3 @@
+export type External = Record<string, string>;
+
+export type ResolvedExternal = Record<string, string>;
