+++ packages/rspack/src/config/mode.ts
@@ -1,0 +1,2 @@
+export type Mode = string;
+export type ResolvedMode = string;
