+++ packages/rspack/src/config/plugin.ts
@@ -1,0 +1,6 @@
+import Rspack from "..";
+
+export interface Plugin {
+	name: string;
+	apply(compiler: Rspack): void;
+}
