+++ packages/rspack/src/config/target.ts
@@ -1,0 +1,2 @@
+export type Target = string;
+export type ResolvedTarget = string;
