+++ packages/rspack/src/index.ts
@@ -1,208 +1,9 @@
 export * from "./build";
-
 import * as binding from "@rspack/binding";
-import type {
-	ExternalObject,
-	RspackInternal,
-	RawModuleRuleUse,
-	RawModuleRule
-} from "@rspack/binding";
+import type { ExternalObject, RspackInternal } from "@rspack/binding";
 import * as tapable from "tapable";
-import assert from "node:assert";
-import * as Config from "./config";
-import type { RspackOptions } from "./config";
-
-interface ModuleRule {
-	test?: RawModuleRule["test"];
-	resource?: RawModuleRule["resource"];
-	resourceQuery?: RawModuleRule["resourceQuery"];
-	uses?: ModuleRuleUse[];
-	type?: RawModuleRule["type"];
-}
-
-type ModuleRuleUse =
-	| {
-			builtinLoader: BuiltinLoader;
-			options?: unknown;
-	  }
-	| {
-			loader: JsLoader;
-			options?: unknown;
-	  };
-
-interface JsLoader {
-	(this: LoaderContext, loaderContext: LoaderContext):
-		| Promise<LoaderResult | void>
-		| LoaderResult
-		| void;
-	displayName?: string;
-}
-
-type BuiltinLoader = string;
-
-interface LoaderThreadsafeContext {
-	id: number;
-	p: LoaderContextInternal;
-}
-
-interface LoaderContextInternal {
-	// TODO: It's not a good way to do this, we should split the `source` into a separate type and avoid using `serde_json`, but it's a temporary solution.
-	source: number[];
-	resource: String;
-	resourcePath: String;
-	resourceQuery: String | null;
-	resourceFragment: String | null;
-}
-
-interface LoaderContext
-	extends Pick<
-		LoaderContextInternal,
-		"resource" | "resourcePath" | "resourceQuery" | "resourceFragment"
-	> {
-	source: {
-		getCode(): string;
-		getBuffer(): Buffer;
-	};
-}
-
-interface LoaderResultInternal {
-	content: number[];
-	meta: number[];
-}
-
-interface LoaderResult {
-	content: Buffer | string;
-	meta: Buffer | string;
-}
-
-interface LoaderThreadsafeResult {
-	id: number;
-	p: LoaderResultInternal | null | undefined;
-}
-
-const toBuffer = (bufLike: string | Buffer): Buffer => {
-	if (Buffer.isBuffer(bufLike)) {
-		return bufLike;
-	} else if (typeof bufLike === "string") {
-		return Buffer.from(bufLike);
-	}
-
-	throw new Error("Buffer or string expected");
-};
-
-function createRawModuleRuleUses(uses: ModuleRuleUse[]): RawModuleRuleUse[] {
-	return createRawModuleRuleUsesImpl([...uses].reverse());
-}
-
-function createRawModuleRuleUsesImpl(
-	uses: ModuleRuleUse[]
-): RawModuleRuleUse[] {
-	const index = uses.findIndex(use => "builtinLoader" in use);
-	if (index < 0) {
-		return [composeJsUse(uses)];
-	}
-
-	const before = uses.slice(0, index);
-	const after = uses.slice(index + 1);
-	return [
-		composeJsUse(before),
-		createNativeUse(uses[index]),
-		...createRawModuleRuleUsesImpl(after)
-	];
-}
-
-function createNativeUse(use: ModuleRuleUse): RawModuleRuleUse {
-	assert("builtinLoader" in use);
-
-	if (use.builtinLoader === "sass-loader") {
-		(use.options ??= {} as any).__exePath = require.resolve(
-			`@tmp-sass-embedded/${process.platform}-${
-				process.arch
-			}/dart-sass-embedded/dart-sass-embedded${
-				process.platform === "win32" ? ".bat" : ""
-			}`
-		);
-	}
-
-	return {
-		builtinLoader: use.builtinLoader,
-		options: JSON.stringify(use.options)
-	};
-}
-
-function composeJsUse(uses: ModuleRuleUse[]): RawModuleRuleUse {
-	async function loader(err: any, data: Buffer): Promise<Buffer> {
-		if (err) {
-			throw err;
-		}
-
-		const loaderThreadsafeContext: LoaderThreadsafeContext = JSON.parse(
-			data.toString("utf-8")
-		);
-
-		const { p: payload, id } = loaderThreadsafeContext;
-
-		const loaderContextInternal: LoaderContextInternal = {
-			source: payload.source,
-			resourcePath: payload.resourcePath,
-			resourceQuery: payload.resourceQuery,
-			resource: payload.resource,
-			resourceFragment: payload.resourceFragment
-		};
-
-		let sourceBuffer = Buffer.from(loaderContextInternal.source);
-		let meta = Buffer.from("");
-		// Loader is executed from right to left
-		for (const use of uses) {
-			assert("loader" in use);
-			const loaderContext = {
-				...loaderContextInternal,
-				source: {
-					getCode(): string {
-						return sourceBuffer.toString("utf-8");
-					},
-					getBuffer(): Buffer {
-						return sourceBuffer;
-					}
-				},
-				getOptions() {
-					return use.options;
-				}
-			};
-
-			let loaderResult: LoaderResult;
-			if (
-				(loaderResult = await Promise.resolve().then(() =>
-					use.loader.apply(loaderContext, [loaderContext])
-				))
-			) {
-				const content = loaderResult.content;
-				meta = meta.length > 0 ? meta : toBuffer(loaderResult.meta);
-				sourceBuffer = toBuffer(content);
-			}
-		}
-
-		const loaderResultPayload: LoaderResultInternal = {
-			content: [...sourceBuffer],
-			meta: [...meta]
-		};
-
-		const loaderThreadsafeResult: LoaderThreadsafeResult = {
-			id: id,
-			p: loaderResultPayload
-		};
-		return Buffer.from(JSON.stringify(loaderThreadsafeResult), "utf-8");
-	}
-	loader.displayName = `NodeLoaderAdapter(${uses
-		.map(item => {
-			assert("loader" in item);
-			return item.loader.displayName || item.loader.name || "unknown-loader";
-		})
-		.join(" -> ")})`;
-	return {
-		loader
-	};
-}
+import * as config from "./config";
+import type { RspackOptions, ResolvedRspackOptions, Assets } from "./config";
 interface RspackThreadsafeContext<T> {
 	readonly id: number;
 	readonly inner: T;
@@ -222,18 +23,20 @@ class Rspack {
 	#instance: ExternalObject<RspackInternal>;
 	#plugins: RspackOptions["plugins"];
 	hooks: {
-		processAssets: tapable.AsyncSeriesHook<Config.Assets>;
+		processAssets: tapable.AsyncSeriesHook<Assets>;
 		done: tapable.AsyncSeriesHook<void>;
 	};
-	constructor(public options: RspackOptions) {
-		const nativeConfig = Config.User2Native(options);
+	options: ResolvedRspackOptions;
+	constructor(options: RspackOptions = {}) {
+		this.options = config.resolveConfig(options);
 
-		this.#instance = binding.newRspack(nativeConfig, {
+		//@ts-ignored TODO: fix it later
+		this.#instance = binding.newRspack(this.options, {
 			doneCallback: this.#done.bind(this),
 			processAssetsCallback: this.#procssAssets.bind(this)
 		});
 		this.hooks = {
-			processAssets: new tapable.AsyncSeriesHook<Config.Assets>(["assets"]),
+			processAssets: new tapable.AsyncSeriesHook<Assets>(["assets"]),
 			done: new tapable.AsyncSeriesHook<void>()
 		};
 		this.#plugins = options.plugins ?? [];
@@ -253,7 +56,7 @@ class Rspack {
 		if (err) {
 			throw err;
 		}
-		const context: RspackThreadsafeContext<Config.Assets> = JSON.parse(value);
+		const context: RspackThreadsafeContext<Assets> = JSON.parse(value);
 		await this.hooks.processAssets.promise(context?.inner);
 		return createDummyResult(context.id);
 	}
@@ -262,12 +65,11 @@ class Rspack {
 		return stats;
 	}
 
-	async rebuild() {
-		const stats = await binding.rebuild(this.#instance);
+	async rebuild(changedFilesPath: string[]) {
+		const stats = await binding.rebuild(this.#instance, changedFilesPath);
 		return stats;
 	}
 }
 
-export { Rspack, createRawModuleRuleUses };
-export type { ModuleRule };
+export { Rspack };
 export default Rspack;
